Refactor newupdater module structure and imports
Moved files from 'newupdater/' to 'newupdater/new_updater/' for improved organization. Updated all internal imports to use the new package structure and relative imports where appropriate. Added an __init__.py to expose key functions and classes.
