+++ newupdater/new_updater/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+"""
+
+from .helps import ec_de_code, print_s
+from .MedWorkNew import work_on_text
+from .mv_section import move_External_links_section
+
+__all__ = [
+    "ec_de_code",
+    "print_s",
+    "work_on_text",
+    "move_External_links_section",
+]
