+++ newupdater/new_updater/bots/expend_new.py
@@ -1,10 +1,10 @@
 """
-from newupdater.bots.expend_new import expend_infoboxs
+from .bots.expend_new import expend_infoboxs
 """
 import wikitextparser as wtp
-from newupdater.helps import echo_debug
+from new_updater.helps import echo_debug
 
-from newupdater.lists.expend_lists import main_temps_list
+from new_updater.lists.expend_lists import main_temps_list
 
 
 def expend_new(template, min_len=1):
