+++ newupdater/new_updater/lists/bot_params.py
